Move the remaining query string types out of Internal namespace
