+++ src/JsonApiDotNetCore/QueryStrings/QueryStringParameterReader.cs
@@ -4,7 +4,7 @@ using JsonApiDotNetCore.Queries.Expressions;
 using JsonApiDotNetCore.Queries.Parsing;
 using JsonApiDotNetCore.Resources.Annotations;
 
-namespace JsonApiDotNetCore.QueryStrings.Internal;
+namespace JsonApiDotNetCore.QueryStrings;
 
 public abstract class QueryStringParameterReader
 {
