+++ src/JsonApiDotNetCore/QueryStrings/SortQueryStringParameterReader.cs
@@ -9,7 +9,7 @@ using JsonApiDotNetCore.Queries.Parsing;
 using JsonApiDotNetCore.QueryStrings.FieldChains;
 using Microsoft.Extensions.Primitives;
 
-namespace JsonApiDotNetCore.QueryStrings.Internal;
+namespace JsonApiDotNetCore.QueryStrings;
 
 /// <inheritdoc cref="ISortQueryStringParameterReader" />
 [PublicAPI]
