+++ src/JsonApiDotNetCore/QueryStrings/SparseFieldSetQueryStringParameterReader.cs
@@ -11,7 +11,7 @@ using JsonApiDotNetCore.Resources;
 using JsonApiDotNetCore.Resources.Annotations;
 using Microsoft.Extensions.Primitives;
 
-namespace JsonApiDotNetCore.QueryStrings.Internal;
+namespace JsonApiDotNetCore.QueryStrings;
 
 /// <inheritdoc cref="ISparseFieldSetQueryStringParameterReader" />
 [PublicAPI]
