+++ test/JsonApiDotNetCoreTests/UnitTests/QueryStringParameters/FilterValueConversionTests.cs
@@ -5,7 +5,6 @@ using JsonApiDotNetCore;
 using JsonApiDotNetCore.Errors;
 using JsonApiDotNetCore.Queries.Parsing;
 using JsonApiDotNetCore.QueryStrings;
-using JsonApiDotNetCore.QueryStrings.Internal;
 using JsonApiDotNetCore.Resources;
 using JsonApiDotNetCore.Resources.Annotations;
 using JsonApiDotNetCore.Serialization.Objects;
